third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/non-anchored-dialog-positioning.html

<!DOCTYPE html>
<html>
<head>
<style>
.filler {
    height: 20000px;
}

dialog {
    /* Remove border and padding to allow comparing dialog's position with that of plain span elements. */
    border: 0;
    padding: 0;
}
</style>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.internals)
    internals.settings.setDialogElementEnabled(true);

function checkTopOfViewport(dialog) {
    shouldBe('dialog.getBoundingClientRect().top', '0');
}

function checkCentered(dialog) {
    expectedTop = (window.innerHeight - dialog.offsetHeight) / 2;
    shouldBe('dialog.getBoundingClientRect().top', 'expectedTop');
}

// Helper to test both a non-modal and modal dialog.
function showAndTestDialog(dialog, checker) {
    dialog.show();
    checker();
    dialog.close();

    dialog.showModal();
    checker();
    dialog.close();
}
</script>
</head>
<body>
    <dialog id="mydialog">It is my dialog.</dialog>
    <div class="filler" id="fillerDiv"></div>
</body>
<script>
description("Tests default positioning of non-anchored dialogs.");

dialog = document.getElementById('mydialog');

debug('Dialog should be centered in the viewport.');
showAndTestDialog(dialog, function() { checkCentered(dialog) });

debug('<br>Dialog should be recentered if show() is called after close().');
window.scroll(0, 500);
dialog.show();
checkCentered(dialog);

debug('<br>Dialog should not be recentered on a relayout.');
var expectedTop = dialog.getBoundingClientRect().top;
window.scroll(0, 1000);
var forceRelayoutDiv = document.createElement('div');
forceRelayoutDiv.style.width = '100px';
forceRelayoutDiv.style.height = '100px';
forceRelayoutDiv.style.border = 'solid';
document.body.appendChild(forceRelayoutDiv);
window.scroll(0, 500);
shouldBe('dialog.getBoundingClientRect().top', 'expectedTop');
document.body.removeChild(forceRelayoutDiv);
dialog.close();

debug('<br>A tall dialog should be positioned at the top of the viewport.');
var dialogInner = document.createElement('div');
dialogInner.className = 'filler';
dialog.appendChild(dialogInner);
showAndTestDialog(dialog, function() { checkTopOfViewport(dialog) });
dialog.removeChild(dialogInner);

debug('<br>The dialog should be centered regardless of the presence of a horizontal scrollbar.');
var fillerDiv = document.getElementById('fillerDiv');
fillerDiv.style.width = '4000px';
showAndTestDialog(dialog, function() { checkCentered(dialog) });
fillerDiv.style.width = 'auto';

debug('<br>Test that centering works when dialog is inside positioned containers.');
var absoluteContainer = document.createElement('div');
absoluteContainer.style.position = 'absolute';
absoluteContainer.style.top = '800px;'
absoluteContainer.style.height = '50px;'
absoluteContainer.style.width = '90%';
dialog.parentNode.removeChild(dialog);
document.body.appendChild(absoluteContainer);
absoluteContainer.appendChild(dialog);
showAndTestDialog(dialog, function() { checkCentered(dialog) });
absoluteContainer.removeChild(dialog);

var relativeContainer = document.createElement('div');
relativeContainer.style.position = 'relative';
relativeContainer.style.top = '20px';
relativeContainer.style.height = '30px';
absoluteContainer.appendChild(relativeContainer);
relativeContainer.appendChild(dialog);
dialog.show();
checkCentered(dialog);

debug("<br>Dialog's position should survive after becoming display:none temporarily.");
expectedTop = dialog.getBoundingClientRect().top;
window.scroll(0, 1000);
relativeContainer.style.display = 'none';
relativeContainer.style.display = 'block';
window.scroll(0, 500);
shouldBe('dialog.getBoundingClientRect().top', 'expectedTop');

debug("<br>Dialog's position should survive after being re-added to document without another call to show().");
expectedTop = dialog.getBoundingClientRect().top;
window.scroll(0, 1000);
relativeContainer.removeChild(dialog);
relativeContainer.appendChild(dialog);
window.scroll(0, 500);
shouldBe('dialog.getBoundingClientRect().top', 'expectedTop');
dialog.close();

debug("<br>Dialog's position should survive after close() and show().");
dialog.show();
dialog.style.top = '0px';
dialog.close();
dialog.show();
var expectedTop = '0px';
shouldBe('dialog.style.top', 'expectedTop');
dialog.style.top = 'auto';

debug("<br>Dialog is recentered if show() is called after removing 'open'");
dialog.removeAttribute('open');
window.scroll(0, 1000);
dialog.show();
checkCentered(dialog);
dialog.close();
window.scroll(0, 500);

debug("<br>Dialog should not be centered if show() was called when an ancestor had display 'none'.");
var plainSpan = document.createElement('span');
plainSpan.style.position = 'absolute';
relativeContainer.appendChild(plainSpan);
expectedTop = plainSpan.getBoundingClientRect().top;
absoluteContainer.style.display = 'none';
dialog.show();
absoluteContainer.style.display = 'block';
shouldBe('dialog.getBoundingClientRect().top', 'expectedTop');
dialog.close();

debug("<br>Test that setting 'top' on dialog results in the same position as for a plain, absolutely positioned span.");
plainSpan.style.top = '50px';
dialog.style.top = '50px';
showAndTestDialog(dialog, function() { shouldBe('dialog.offsetTop', 'plainSpan.offsetTop'); });

debug("<br>Test that setting 'bottom' on dialog results in the same position as for a plain, absolutely positioned span.");
dialog.style.top = 'auto';
plainSpan.style.top = 'auto';
dialog.style.bottom = '50px';
plainSpan.style.bottom = '50px';
showAndTestDialog(dialog, function() { shouldBe('dialog.offsetBottom', 'plainSpan.offsetBottom'); });

debug('<br>Test that fixed positioning for dialog has same behavior as for a plain span.');
plainSpan.style.position = 'fixed';
plainSpan.style.top = '50px';
dialog.style.position = 'fixed';
dialog.style.top = '50px';
showAndTestDialog(dialog, function() { shouldBe('dialog.offsetTop', 'plainSpan.offsetTop'); });

debug('<br>Test that static position for a non-modal dialog has the same behavior as for a plain span.');
// Just test non-modal since modal is covered by other tests (for modal, static computes to absolute)
plainSpan.style.position = 'static';
var expectedTop = plainSpan.offsetTop;
plainSpan.parentNode.removeChild(plainSpan);
dialog.style.position = 'static';
dialog.show();
shouldBe('dialog.offsetTop', 'expectedTop');
dialog.close();

debug('<br>Test that relative position for a non-modal dialog has the same behavior as for a plain span.');
// Just test non-modal since modal is covered by other tests (for modal, relative computes to absolute)
relativeContainer.appendChild(plainSpan);
plainSpan.style.position = 'relative';
plainSpan.style.top = '50px';
expectedTop = plainSpan.offsetTop;
plainSpan.parentNode.removeChild(plainSpan);
dialog.style.position = 'relative';
dialog.style.top = '50px';
dialog.show();
shouldBe('dialog.offsetTop', 'expectedTop');
dialog.close();
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
